Release 10.35.88
git-svn-id: http://svn.code.sf.net/p/netpbm/code/super_stable@1853 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/pbm/cmuwmtopbm.c
@@ -78,7 +78,7 @@ getinit( file, colsP, rowsP, depthP, padrightP )
 
     if ( pm_readbiglong( file, &l ) == -1 )
 	pm_error( "EOF / read error" );
-    if ( l != CMUWM_MAGIC )
+    if ( (uint32_t)l != CMUWM_MAGIC )
 	pm_error( "bad magic number in CMU window manager file" );
     if ( pm_readbiglong( file, &l ) == -1 )
 	pm_error( "EOF / read error" );
